Extract OptionSchema to another file
